+++ app/models/design.rb
@@ -368,7 +368,8 @@ class Design < ApplicationRecord
 
   # Reset all associated sheets total_response_count to nil to trigger refresh of sheet answer coverage
   def reset_sheet_total_response_count
-    sheets.update_all(total_response_count: nil, percent: nil)
+    sheets.where(missing: false).update_all(response_count: nil, total_response_count: nil, percent: nil)
+    sheets.where(missing: true).update_all(response_count: 0, total_response_count: 0, percent: 100)
   end
 
   def set_slug